refactor: assets 폴더 이미지 파일명 수정
