+++ src/components/Card/StarIcon.jsx
@@ -1,4 +1,5 @@
 import StarImg from '../../assets/star.svg';
+
 function StarIcon() {
   return <img className='content__star' src={StarImg} alt='별 아이콘' />;
 }
