+++ src/components/modal/ModalCategoryList.jsx
@@ -1,4 +1,5 @@
-import ModalCheckIcon from '../../assets/Modal-check.svg';
+import ModalCheckIcon from '../../assets/modal-check.svg';
+
 import { useState } from 'react';
 function ModalCategoryList({ categoryListLoop }) {
   const [selectCategory, setSelectCategory] = useState(null);
