+++ src/components/modal/ModalCloseIcon.jsx
@@ -1,4 +1,5 @@
 import ModalCloseIcon from '../../assets/modal-close.svg';
+
 function ModalCloseButton({ onClick }) {
   return <img onClick={onClick} className='modal__close' src={ModalCloseIcon} alt='모달 닫기 아이콘' />;
 }
