+++ src/components/modal/ModalShareList.jsx
@@ -1,6 +1,7 @@
 import ModalKaKaoIcon from '../../assets/modal-kakao.svg';
 import ModalFacebookIcon from '../../assets/modal-facebook.svg';
 import ModalLinkIcon from '../../assets/modal-link.svg';
+
 const modalShareList = [
   {
     id: 1,
